feat(game.web): implement a web version of game engine with a landing page
