+++ Game/index-rnsge.tsx
@@ -1,0 +1,13 @@
+import { FC } from 'react';
+import { MainScene } from '../src/containers/Scenes/MainScene';
+import { ReactNativeSkiaGameEngine } from '../src/containers/ReactNativeSkiaGameEngine';
+
+export const Game: FC<unknown> = () => {
+  return (
+    <ReactNativeSkiaGameEngine onEventListeners={{}}>
+      <MainScene />
+    </ReactNativeSkiaGameEngine>
+  );
+};
+
+export default Game;
