+++ Game/loadGame.tsx
@@ -1,0 +1,7 @@
+import React, { FC } from 'react';
+import { Text } from 'react-native';
+import { WithSkiaWeb } from '@shopify/react-native-skia/lib/module/web';
+import { Game } from './index-rnsge';
+export const LoadGame: FC = () => {
+  return <Game />;
+};
