+++ app.config.js
@@ -31,6 +31,7 @@ export default ({ config }) => {
       },
       web: {
         favicon: './assets/favicon.png',
+        bundler: 'metro',
       },
       experiments: {
         tsconfigPaths: true,
@@ -41,6 +42,7 @@ export default ({ config }) => {
         },
         storybookEnabled: process.env.STORYBOOK_ENABLED,
       },
+      plugins: ['expo-font'],
     },
   };
 };
