+++ metro.config.js
@@ -1,14 +1,15 @@
-const { getDefaultConfig } = require("expo/metro-config");
-const path = require("path");
+const { getDefaultConfig } = require('expo/metro-config');
+const { withNativeWind } = require('nativewind/metro');
+const path = require('path');
 
-const { generate } = require("@storybook/react-native/scripts/generate");
+const { generate } = require('@storybook/react-native/scripts/generate');
 
 generate({
-  configPath: path.resolve(__dirname, "./.storybook"),
+  configPath: path.resolve(__dirname, './.storybook'),
 });
 
 const defaultConfig = getDefaultConfig(__dirname);
 
 defaultConfig.transformer.unstable_allowRequireContext = true;
 
-module.exports = defaultConfig;
+module.exports = withNativeWind(defaultConfig, { input: './global.css' });
