+++ package.json
@@ -1,7 +1,7 @@
 {
   "name": "reactnativegame",
   "version": "1.0.0",
-  "main": "node_modules/expo/AppEntry.js",
+  "main": "index",
   "scripts": {
     "start": "expo start",
     "android": "expo start --android",
@@ -14,6 +14,7 @@
     "test": "jest --watchAll"
   },
   "dependencies": {
+    "@expo/metro-runtime": "~4.0.1",
     "@react-native-async-storage/async-storage": "1.23.1",
     "@react-native-community/datetimepicker": "8.2.0",
     "@shopify/react-native-skia": "1.5.0",
@@ -21,11 +22,14 @@
     "cross-env": "^7.0.3",
     "expo": "^52.0.19",
     "expo-dev-client": "~5.0.6",
-    "expo-linear-gradient": "~14.0.1",
+    "expo-font": "~13.0.4",
+    "expo-linear-gradient": "~14.0.2",
+    "expo-splash-screen": "~0.29.22",
     "expo-status-bar": "~2.0.0",
     "jest-expo": "~52.0.2",
     "lodash": "^4.17.21",
     "matter-js": "^0.18.0",
+    "nativewind": "^4.1.23",
     "punycode": "^2.3.1",
     "react": "18.3.1",
     "react-dom": "18.3.1",
@@ -34,7 +38,9 @@
     "react-native-gesture-handler": "~2.20.2",
     "react-native-reanimated": "~3.16.1",
     "react-native-safe-area-context": "4.12.0",
+    "react-native-web": "~0.19.13",
     "react-test-renderer": "18.2.0",
+    "tailwindcss": "^3.4.17",
     "uuid": "^10.0.0",
     "zustand": "^4.5.2"
   },
