+++ src/components/ShipView/ShipView-rnsge.tsx
@@ -121,6 +121,7 @@ export const ShipView: FC<IShipViewProps> = ({ seaEntityId }) => {
       { translateY: (sharedBody?.value.position.y || 0) - initialY },
     ];
   }, [sharedBody]);
+  console.log('🚀 ~ transform:', transform);
 
   const shipSystem = useMemo(() => new ShipSystem(), []);
   const systemCallback: system = useCallback(
@@ -202,7 +203,11 @@ export const ShipView: FC<IShipViewProps> = ({ seaEntityId }) => {
     ];
   }, [frameWidth, frameHeight, frameIndex]);
 
-  const boatImage = useImage(require('../../../assets/fisher_boat3.png'));
+  const url = require('../../../assets/fisher_boat3.png');
+  console.log('🚀 ~ url:', url);
+
+  const boatImage = useImage(url?.uri);
+  console.log('🚀 ~ boatImage:', boatImage);
   if (!boatImage) return null;
 
   return (
