+++ src/containers/Overlay/index.tsx
@@ -1,8 +1,8 @@
-import { StyleSheet, Text, View, useWindowDimensions } from "react-native";
-import React, { FC, useRef } from "react";
-import Game from "../Game";
-import StartButton from "../../components/UI/StartButton";
-import { useGameState } from "../../store/useGameState";
+import { StyleSheet, Text, View, useWindowDimensions } from 'react-native';
+import React, { FC, useRef } from 'react';
+import Game from '../../../Game';
+import StartButton from '../../components/UI/StartButton';
+import { useGameState } from '../../store/useGameState';
 
 const Overlay: FC<{}> = () => {
   const { width, height } = useWindowDimensions();
@@ -11,10 +11,10 @@ const Overlay: FC<{}> = () => {
   const onStartPress = () => {
     if (isGameRunning) {
       stopGame();
-      gameRef.current.gameEngineRef.current.dispatch("gameOver");
+      gameRef.current.gameEngineRef.current.dispatch('gameOver');
     } else {
       startGame();
-      gameRef.current.gameEngineRef.current.dispatch("startGame");
+      gameRef.current.gameEngineRef.current.dispatch('startGame');
     }
   };
   return (
@@ -43,19 +43,19 @@ export default Overlay;
 const styles = StyleSheet.create({
   root: {
     flex: 1,
-    justifyContent: "center",
-    alignItems: "center",
-    backgroundColor: "green",
+    justifyContent: 'center',
+    alignItems: 'center',
+    backgroundColor: 'green',
   },
   overlayRoot: {
-    position: "absolute",
-    backgroundColor: "transparent",
+    position: 'absolute',
+    backgroundColor: 'transparent',
     zIndex: 2,
   },
   gameRoot: {
-    position: "absolute",
-    width: "100%",
-    height: "100%",
+    position: 'absolute',
+    width: '100%',
+    height: '100%',
     // zIndex: 4,
     top: 0,
     left: 0,
